article test

+++ internal/web/middleware/login_jwt.go
@@ -65,7 +65,7 @@ func (l *LoginJWTMiddlewareBuilder) Build() gin.HandlerFunc {
 		//}
 		//ctx.Header("x-jwt-token", tokenStr)
 		ctx.Set("claims", claims)
-		ctx.Set("userId", claims.Uid)
+		ctx.Set("userId", int64(claims.Uid))
 	}
 
 }
